requirements.txt
alembic==1.9.3
altair==4.2.2
aniso8601==9.0.1
aspose-words==23.1.0
attrs==22.2.0
azure-ai-formrecognizer==3.2.0
azure-common==1.1.28
azure-core==1.26.3
backoff==2.2.1
backports.zoneinfo==0.2.1
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
cryptography==39.0.1
databricks-cli==0.17.4
decorator==5.1.1
dill==0.3.6
docker==6.0.1
docopt==0.6.2
easyocr==1.6.2
elasticsearch==7.17.9
entrypoints==0.4
farm-haystack==1.13.1
filelock==3.9.0
Flask==2.2.2
Flask-RESTful==0.3.9
gitdb==4.0.10
GitPython==3.1.30
greenlet==2.0.2
gunicorn==20.1.0
huggingface-hub==0.12.0
idna==3.4
imageio==2.25.0
importlib-metadata==5.2.0
importlib-resources==5.10.2
inflect==6.0.2
isodate==0.6.1
itsdangerous==2.1.2
jarowinkler==1.2.3
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
langdetect==1.0.9
lxml==4.9.2
Mako==1.2.4
markdown-it-py==2.1.0
MarkupSafe==2.1.2
mdurl==0.1.2
mlflow==1.30.0
mmh3==3.0.0
monotonic==1.6
more-itertools==9.0.0
msrest==0.7.1
mysql-connector-python==8.0.32
networkx==2.6.3
ninja==1.11.1
nltk==3.8.1
num2words==0.5.12
numpy==1.21.6
oauthlib==3.2.2
opencv-python-headless==4.5.4.60
packaging==21.3
pandas==1.3.5
pdf2image==1.16.2
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
posthog==2.3.1
prometheus-client==0.16.0
prometheus-flask-exporter==0.21.0
protobuf==3.20.2
pyarrow==11.0.0
pyclipper==1.3.0.post4
pycparser==2.21
pydantic==1.10.4
pydeck==0.8.0
Pygments==2.14.0
PyJWT==2.6.0
Pympler==1.0.1
pyodbc==4.0.35
pyparsing==3.0.9
PyPDF2==2.11.1
pyrsistent==0.19.3
python-bidi==0.4.2
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
pywin32==305
PyYAML==6.0
quantulum3==0.8.1
querystring-parser==1.2.4
rank-bm25==0.2.2
rapidfuzz==2.7.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.1
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
semver==2.13.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
seqeval==1.2.2
shapely==2.0.1
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.46
sqlparse==0.4.3
streamlit==1.17.0
tabulate==0.9.0
threadpoolctl==3.1.0
tifffile==2021.11.2
tika==2.6.0
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
transformers==4.25.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
waitress==2.1.2
watchdog==2.2.1
websocket-client==1.5.1
Werkzeug==2.2.2
zipp==3.12.1
